+++ app/lib/features/activities/widgets/space_activities_section/item_widgets/space_activities_item_widget.dart
@@ -7,6 +7,7 @@ import 'package:atlas_icons/atlas_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:skeletonizer/skeletonizer.dart';
+import 'package:acter/common/widgets/dashed_line_vertical.dart';
 
 class SpaceActivitiesItemWidget extends ConsumerWidget {
   const SpaceActivitiesItemWidget({
@@ -97,65 +98,4 @@ class SpaceActivitiesSkeleton extends StatelessWidget {
       ),
     );
   }
-}
-
-class DashedLineVertical extends StatelessWidget {
-  final double? height;
-  final double dashHeight;
-  final double dashSpacing;
-  final Color color;
-
-  const DashedLineVertical({
-    super.key,
-    this.height,
-    this.dashHeight = 6,
-    this.dashSpacing = 5,
-    this.color = Colors.grey,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 0.5,
-      height: height,
-      child: CustomPaint(
-        painter: _DashedLinePainter(
-          dashHeight: dashHeight,
-          dashSpacing: dashSpacing,
-          color: color,
-        ),
-      ),
-    );
-  }
-}
-
-class _DashedLinePainter extends CustomPainter {
-  final double dashHeight;
-  final double dashSpacing;
-  final Color color;
-
-  _DashedLinePainter({
-    required this.dashHeight,
-    required this.dashSpacing,
-    required this.color,
-  });
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    double startY = 0;
-    final paint = Paint()
-      ..color = color
-      ..strokeWidth = 0.8;
-    while (startY < size.height) {
-      canvas.drawLine(
-        Offset(size.width / 2, startY),
-        Offset(size.width / 2, startY + dashHeight),
-        paint,
-      );
-      startY += dashHeight + dashSpacing;
-    }
-  }
-
-  @override
-  bool shouldRepaint(CustomPainter oldDelegate) => false;
 }
